% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{searchgbif}
\alias{searchgbif}
\title{Search for GBIF occurrences.}
\usage{
searchgbif(scientificName = NULL, limit = 20, cache = FALSE, db = NULL,
  backend = "local", path = "~/")
}
\arguments{
\item{scientificName}{Scientific name}

\item{limit}{Number of results to return}

\item{cache}{Logical, defaults to FALSE}

\item{db}{Database}

\item{backend}{One of local, redis, couchdb, mongodb, sqlite, git.}

\item{path}{Path to a directory on your machine.}
}
\description{
Search for GBIF occurrences.
}
\examples{
\dontrun{
# Don't cache
searchgbif(scientificName='Ursus americanus', limit=2)

# Cache locally, using saveRDS
searchgbif(scientificName='Ursus americanus', limit=2, cache=TRUE, backend='local')

# Cache using redis
searchgbif(scientificName='Ursus americanus', limit=2, cache=TRUE, backend='redis')

# Cache using rcache, via the R.cache package
searchgbif(scientificName='Ursus americanus', limit=2, cache=TRUE, backend='rcache')

# Optionally, you can set mocker as options, then mocker looks for them
# This makes your function calls cleaner by avoiding the extra parameter settings
# If you use options, arguments passed in to the function call will override options settings

options(cache=TRUE)
options(backend='redis')
searchgbif(scientificName='Ursus', limit=2) # uses stored options
searchgbif(scientificName='Ursus', limit=2, cache=FALSE) # override cache option setting
}
}

